Squirrel function default arguments support
This PR updates `simplesquirrel` to a branch which supports exposing C++ functions (including class constructors) wtih default arguments. It's not yet been merged into simplesquirrel `master`, because the Squirrel branch used is also not `master`, so an important change to Squirrel parameter checking is included to allow for default argument support.

Only the `Camera::scale()` and `scripting::Level::spawn()` functions were modified to support default arguments and their respective other functions were marked as deprecated. In case more functions are found suitable for this, this PR would be updated.