Support "info" files in add-on archives, enable/disable add-ons in config
Add-on archive "info" files contain the following info about the add-on: title, author, whether it should override data (be prepended to the tree), whether it is considered a resource pack and optionally, any other add-ons it has as dependencies.

Enabling/disabling add-ons is planned to be possible via an "Add-ons" menu in-game. For now, doing so is only possible through the config and will only be applied on game startup.