`MusicResource`: Always free `Mix_Music*` and `SDL_RWops*`; do not utilize a separate `MusicManager` for credits
The `SDL_RWops` pointer that was passed to `Mix_LoadMUS_RW` has to manually be freed, as SDL_mixer 1.x does not automatically do that. This fixes an issue preventing add-ons with custom music from being unmounted after any of their music was loaded once.

Additionally, the credits no longer use a separate `MusicManager` to play music. This eliminates the annoying lag when quitting the credits.